third_party/WebKit/LayoutTests/inspector/styles/stylesheet-source-url-comment-expected.txt

Tests that stylesheets with sourceURL comment are shown in the Sources panel.


Running: testSourceURLCommentInInlineScript

Running: testSourceURLComment
body { color: black; }

Running: testDeprecatedSourceURLComment
body { color: black; }

Running: testNonRelativeURL
body { color: red; }
